+++ Modules/core/Controller/CorespaceController.php
@@ -81,6 +81,7 @@ class CorespaceController extends CoresecureController {
         $lang = $this->getLanguage();
         $role = 0;
         $showAdmMenu = false;
+        $isMemberOfSpace = false;
 
         if ($_SESSION['user_status'] > CoreStatus::$USER) {
             $showAdmMenu = true;
@@ -89,6 +90,8 @@ class CorespaceController extends CoresecureController {
             $role = $this->spaceModel->getUserSpaceRole($space["id"], $_SESSION['id_user']);
             if ($role == -1) {
                 $role = CoreSpace::$VISITOR;
+            } else {
+                $isMemberOfSpace = true;
             }
             if ($role > CoreSpace::$MANAGER) {
                 $showAdmMenu = true;
@@ -128,6 +131,7 @@ class CorespaceController extends CoresecureController {
         }
         return $this->render(array(
             "role" => $role,
+            "isMemberOfSpace" => $isMemberOfSpace,
             "lang" => $lang,
             "id_space" => $id_space,
             "space" => $space,